fix: resolve frontend security alerts
