+++ frontend/package.json
@@ -51,7 +51,7 @@
     "postcss": "^8.5.6",
     "tailwindcss": "^4.1.18",
     "typescript": "~5.9.3",
-    "vite": "^7.2.4",
+    "vite": "^7.3.2",
     "vitest": "^4.1.2",
     "vue-tsc": "^3.1.4"
   }
